Make --noallow_discard_analysis_cache work when used twice in a seque… (#23653)
…nce.

This didn't work at HEAD because the check there compared the flags in
the new configuration with the flags of the old configuration, but
overwrote the old configuration with the new one, so running a command
that discarded the analysis cache was allowed the second time.

The fix is not as trivial as moving checking the value before setting
the new configuration because creating the new configuration entails a
Skyframe evaluation, which is a waste if we end up erroring out. So the
check is now done by checking the build options of the old configuration
and the new build options, which doesn't require creating the new
configuration first.

Fixes #23491.

RELNOTES: None.
PiperOrigin-RevId: 675555324
Change-Id: Ib402b699ed9ba9f04542896a40ff1180351471db