Apply length check on the user-defined attribute name for subrule attributes
Subrule attributes are lifted to the using rule class. As currently implemented, for an attribute named `a` in a subrule `sub` defined in `//path/to:ext.bzl`, the corresponding attribute name in the rule class is `$//path/to:ext.bzl%sub%a`. Since labels can be fairly long, this may exceed the limit on attribute names.

PiperOrigin-RevId: 675995599
Change-Id: Ied8501b5d34d2373ca3625d6115a9ac93c4ec8b4