Add macro visibility delegation
This CL adds another piece of the Macro-Aware Visibility model: that a symbolic macro may delegate its visibility privileges over a dependency target to a submacro that would not ordinarily be able to see the dependency. This is needed to ensure that a dependency target does not need to modify its visibility allowlist to account for the implementation details of all the macros it is passed to. It also eliminates a source of breakages when refactoring a rule into a symbolic macro.

The semantics are simply this: If you have a dependency on a target, and you were declared inside a (symbolic) macro that took this dependency's label in one of its attributes, then the visibility check applies with respect to your containing macro instead of with respect to you.

- If the macro can see it, then then macro's considered to have delegated this privilege to you, so it's irrelevant whether you can see it or not.

- On the other hand, if then macro can't see it, then it's an error that the macro's caller tried to pass it to the macro in the first place, so we fail regardless of whether or not you yourself were allowed to see it.

Note that this process is recursive; the macro may be able to see the dependency by virtue of an outer macro that passed the dependency's label in. (The implementation of the check recurses *up* to the parent, while we say that the parent delegated its privilege *down* to the child.)

In some cases a macro can have an illegal dependency that does not cause a build failure because this usage did not make it into any built target (including via delegation to a submacro).

Changes:
- `CommonPrerequisiteValidator` implements the heart of the new delegation behavior. `isVisibleToRule()` is broadened to `isVisibleToDeclaration()`, so it can recurse to macro declarations and not just handle target declarations.
- Added `MacroInstance#visitExplicitAttributeLabels()` to gather labels in macro attributes (excluding implicit attributes, and excluding output labels and nondep labels like the future `visibility` attribute). This is complicated by the fact that macros store attribute values as Starlark values rather than their internal equivalent (the way rules do it). But we are able to leverage existing logic in `AggregatingAttributeMapper` for traversing selects (made a static method there public).
- Add clarifying comments that would've helped me while I was browsing this code.
- Another half a kLOC of test cases in `MacroVisibilityTest`. Broadened the assertion machinery to be able to test `alias` targets, which produce slightly different visibility error messages. Saved 50 LOC and some readability by factoring creation of common/BUILD into a helper.
- Add tests to `SymbolicMacroTest` that None can be used in selects for macros (like they can be for rules), and of new label visitation method.
- Delete `ObjectType` and `ObjectListType` subclasses of `Type`. These are unused, unusable, and date back to primordial code from 2006. I'm amazed they've lasted as zombie code for this long.

Work toward #19922.

PiperOrigin-RevId: 676092563
Change-Id: I1f8098f0bbb9de3566b9693b7db5a5c7cdac942f