Reorder the top-level declarations in the Package class
The declaration order of the elements in `Package.java` makes it even harder to figure out what is going on in this already monstrous (3k LOC) file. That order evolved over a very long time. This CL is an attempt to wrangle this complexity and unblock further cleanups, toward the goal of adding support for evaluation of individual symbolic macros.

(The drawback of this change is that it has a very large blast radius in terms of code line history / blame. I imagine I'll be sent many CL reviews by default for Package.java in the future...)

The new order for `Package`'s members is:

1. enums and static fields
2. instance fields -- metadata, then WORKSPACE-like fields, then targets/macros maps
3. accessors, in the same order as their corresponding fields (or where those fields will be after the follow-up CL). This includes a number of convenience accessors around subfields of the `metadata` field.
4. initialization logic - constructor, `finishInit()`, and a mutator only used in construction (I hope)
5. stringification
6. public static utility methods
7. non-trivial subclasses

with private helper methods sprinkled where appropriate. The order for `Package.Builder` and `Package.Metadata` elements is unchanged.

Other non-ordering changes:
- inlined `setPackageOverhead()`
- `getRepositoryMapping(RepositoryName)` was an overload of `getRepositoryMapping()`, but the former is intended for use only in `//external` while the latter is for normal `BUILD` packages. Renamed `getRepositoryMapping(RepositoryName)` -> `getExternalPackageRepositoryMapping(RepositoryName)`, which almost overloads an analogous method `getExternalPackageRepositoryMappings()` (note the missing 's').
- elaborated TODO for `setContainsErrors()`
- made `contextualizeFailureDetailForTarget` a static method, eliminated unneeded preconditions
- renamed static overload of `getSourceRoot` to `computeSourceRoot`; it belongs with `finishInit`, not the accessor
- inlined `createDetailedCode`

Work toward #19922.

PiperOrigin-RevId: 676094017
Change-Id: If2f779a6461b45c9c586653ee31f2f9dda7b71eb