[7.4.0] Allow unquoted canonical repository names with `query` (#23675)
Unquoted `query` words that start with `@@` are no longer broken on `+`,
which allows the usage of unquoted canonical labels in expressions. In
the very rare case of a shorthand form of a canonical label referring to
a WORKSPACE repo (which contains no `+`), the `+` can be separated by a
space to get the original behavior (e.g. `@@foo+bar` no longer means
`@@foo//:foo + //bar:bar`, but `@@foo +bar` does.

Closes #23600.

PiperOrigin-RevId: 673763980
Change-Id: Ie5509647b4e04e2f72b908d609781c79cf7b06d4

Fixes #23601