Push Metadata out of Builder constructor
Now the three builder-factor methods construct the `Metadata` and pass it as a constructor arg. This simplifies the constructor signature and helps with passing metadata to a super() constructor in a follow-up CL.

The `isRepoRulePackage` calculation is only done in one of the constructors, and there it's only needed to work around a hack in package deserialization that I don't want to address right now (and which may be obsolete after WORKSPACE logic is deleted).

Work toward #19922.

PiperOrigin-RevId: 676530459
Change-Id: Ieaabeb412925886a0452447f186a9bb15a3a21c5