[7.4.0] Strip unnecessary Bazel server jar contents
fastutil is very large and thus stripped (but not obfuscated) with proguard. The Auto* processors do not need to be included into the release binary, only their annotations.

Also removes some third_party targets that are unused.

Bazel binary size on macOS is 65 MB before and 53 MB after this change.

Work towards #22722