Automated rollback of commit 9acc146bfae04e9071cc73331cb667fdd2bb9fcb.
*** Reason for rollback ***

Pivoting to a different approach.

*** Original change description ***

Add a starting point for the disk cache garbage collector.

This is not yet a fully functional implementation. For now, the garbage collector (when enabled by the --experimental_disk_cache_max_size flag) merely checks whether the index database is missing or empty and, if so, recreates it from the filesystem contents.

PiperOrigin-RevId: 677739324
Change-Id: Ic80419c01cf35fce034350b60ea6a8ebb6ae605b