[7.4.0] Fix `runtime_solib_name` for `--incompatible_macos_set_install_name` (#23672)
The `runtime_solib_name` link variable had an incorrect value in the
case of a transitive dynamic library. Since non-transitive ("nodeps")
dynamic libraries are no longer used on macOS, the
`--incompatible_macos_set_install_name` flag didn't have any positive
effect.

This PR is intentionally limited to the fix so that it can be
cherry-picked into Bazel 7, where it can make the incompatible flag work
with the `apple_support` toolchain. A follow-up PR will add the feature
to the Unix toolchain and flip the incompatible flag for Bazel 8.

Work towards #12370

Closes #23089.

PiperOrigin-RevId: 668228562
Change-Id: I7524679bfe8c6b8b28c0cb04f46c0d22d0adbe99

Fixes #23185