Add integration test to demonstrate effect of baseline options to the configuration fragment in the artifact output path.
This is likely to be a invalidation problem for deserialization in analysis caching because an invocation can get a cache hit from a build with totally different top level flags. Or in Skyframe lingo: the same BuildConfigurationKey can result in a BuildConfigurationValue containing a different mnemonic, making it dependent on the invocation environment.

PiperOrigin-RevId: 678267366
Change-Id: I638e9f672b8d18cd7043b7934a15cb1791d2b07a