Fix counting created actions to not over count aspect actions on alias targets
if an aspect is applied on an alias, the resultant `AspectValue` holds the `ConfiguredAspect` of the aspect on the alias's actual target. This results in counting the actions created by the aspect multiple times, one time per target in the alias chain. This CL adds a new class `AspectValueForAlias` to distinguish this case and updates the action counting logic to ignore actions for this `SkyValue`.

PiperOrigin-RevId: 678288884
Change-Id: I9428eac955d2026682e5fb93e2ec62d108bb2fd7