[7.4.0] cquery: Factor RepositoryName into fetching ignored packages (#23727)
Previously this code was assuming that all cquery target patterns should
use the root repository's .bazelignore contents which was incorrect.

Because the way this manifested itself was that when cquery drives
populating the graph and when it queries it, it would use different
keys, this led to misleading messages about packages not being in scope.

Closes #22142.

PiperOrigin-RevId: 628477138
Change-Id: I84e3381848b843ca9813acdd0762a3b3e4711a90