Add a new flag, `--allowed_cpu_values`, to verify that values passed to `--cpu` make sense.
Previously, the value used with `--cpu` had to match an entry in the `cc_toolchain_suite` used by the `--crosstool_top` flag. Since this flag and `cc_toolchain_suite` are deprecated, users can pass any value to `--cpu` with no checking.

This change adds a check if `--allowed_cpu_values` is set. If it is not, there is no validation of the `--cpu` flag.

RELNOTES: If `--allowed_cpu_values` is set, the `--cpu` flag value is validated against it.
PiperOrigin-RevId: 678401605
Change-Id: I5496e43c8dcc7a464ac81f9870cf8fb24a78cc20