Make Package.Builder compose instead of inherit from TargetRegistrationEnvironment
This gives the Builder more explicit control over its own API, at the cost of a bit of verbosity in accessing the environment as a field, and in enumerating the methods it chooses to re-export to its own clients.

It also means that in the future, we can repeat the trick of factoring out a bunch of builder methods to a new base class, without having to worry about which base class supersedes the other in the hierarchy.

- `Pakcage.Builder` now inherits directly from `StarlarkThreadContext`.

- None of `TargetRegistrationEnvironment`'s members are protected, and only a few are package-private. The choice of which methods are public now (mostly) comes down to what API makes sense for `TargetRegistrationEnvironment`, without regard for what makes sense for `Package.Builder`'s clients.

- In the builder, add `currentMacro()` convenience method.

Work toward #19922.

PiperOrigin-RevId: 678411561
Change-Id: I5e550fe5b582e1039fa304d42d37e882a856e742