Impl Arena/Bump allocator.
The Arena reserves a large address space ahead of time and commit/decommit pages on demand.

For production usages, one may need to chain Arenas if it runs out of reserved address. But this is only for tests, using a fixed reserved space should be enough and make the implementation easier.

PiperOrigin-RevId: 679117001
Change-Id: Ib7a938bf87532f3086811e28adb8e524fa44d85b