Allow the same output group to be provided twice.
From now on, if a rule and an aspect applied to it (or two aspects) provide the same output group, the files in it will be merged instead of raising an error.

This is needed to land https://github.com/bazelbuild/bazel/pull/23589 : with that change, the validation aspect depends on every other aspect applied to the rule and the rule itself, which sometimes results in output group conflicts.

RELNOTES: None.
PiperOrigin-RevId: 679554441
Change-Id: I6ee4f5522a69c209746119164497a37960d432e2