[7.4.0] Automated rollback of commit 8ef7b705bb33de6ee9025c01bdf65a23f73926e3.
*** Reason for rollback ***

The experiment with using sqlite didn't pan out. We might find other uses for it in the future, but we can always resurrect it from version control; no need to saddle Bazel with the additional dependency until then.

*** Original change description ***

Implement a JNI wrapper around SQLite.

This will be used by the implementation of garbage collection for the disk cache, as discussed in https://github.com/bazelbuild/bazel/issues/5139 and the linked design doc. I judge this to be preferred over https://github.com/xerial/sqlite-jdbc for the following reasons:

1. It's a much smaller dependency.
2. The JDBC API is too generic and becomes awkward to use when dealing with the peculiarities of SQLite.
3. We can (more easily) compile it from source for all host platforms, including the BSDs.

***

PiperOrigin-RevId: 679600756
Change-Id: Ic3748fa30404a31504426c523c9b9a60ec451863