RuleVisibility concatenation doesn't add redundant items
This is useful if you have several nested symbolic macros, since it avoids generating a long visibility value like `["//A:__pkg__", "//A:__pkg__", ..., "//A:__pkg__"]` when each macro's body is defined in the same package `A`.

Also clarify the purpose of the label accessors on `RuleVisibility` and flip their declaration order.

Work toward #19922.

PiperOrigin-RevId: 679962558
Change-Id: Iebb55d0256d72d7bc0a8b575446dea83c9bbf632