Add a `keyBytes` distinguisher for FrontierSerializer and SkyValueRetriever.
This supports adding the fingerprint of the active directories in the
cache key.

This prevents a correctness problem where a deserializing build gets a top-down
cache hit on a remote configured target value outside of the active_directories
boundary, without knowing that it is in its active set, and incorrectly
skipping local analysis. The cache hit could be created from a prior build's
frontier with a different set of active directories.

PiperOrigin-RevId: 680475998
Change-Id: I26342a7e840b20bb872339568d673449c937d77d