Blunt mitigation for blaze crashing with includes_scanning error: Adding retries for the 2 error modes.
The as yet not understood root cause of this crash must be an unhandled race condition where a *.blaze-grepped_includes_CPP file is read before it is present and fully available.
The two fixes are:
1. Waiting for the file to exist if it doesn't. This wait is very cheap, and it covers the majority of error cases.
2. Retry reading the file should an IOError occur while reading the file. This is needed in fewer cases, and is a slightly heavier measure to take. For that reason, and because this is intended as a temporary mitigation, I left the two separate fixes in place, even though technically the 2nd fix could also cover for the first one.

In any case, the investigation for the root cause should continue, and the temporary fix is placed in the code in a way (incl. TODOs) that makes it easy to remove it again.

The effectiveness of the fix was verified with 200+ successful iterations of clean builds of //superroot/servers:sr_www.

PiperOrigin-RevId: 680537404
Change-Id: I46f2bae6e59a69be6d17252a2cb626d1a2e009a0