Incorporate the top level configuration into the `SkyKey` distinguisher.
This will invalidate remote analysis values if the top level configuration (or
baseline options / command line starlark flags) changes with respect to their
delta to transitioned configurations, but also doesn't change the `BuildConfigurationKey` of a target. This is subtle because the delta to the baseline options affect the mnemonic fragment of the output path.

From the Skyframe point of view, this means two `BuildConfigurationValue`s with different outputs for #getMnemonic can have identical `BuildConfigurationKey`s, and this CL prevents the deserialization pass from fetching an incorrect `ConfiguredTargetValue` (whose key contains a `BuildConfigurationKey`).

PiperOrigin-RevId: 680590271
Change-Id: Ia281d0306a1167c0acb395a0d08dc1598d50d1e6