Allow decompression of nupkg
In some (legacy) projects, nuget is used to handle C++ packages.
Nuget packages are working fine with WORKSPACE `http_archive` directives when explicitly specifying `type = "zip"`.

However, when migrating to bzlmod, we ran into issues when the url in source.json is a nupkg file.
This PR resolves that issue and allows consumption of nupkg archives as Bazel modules.

Closes #23645.

PiperOrigin-RevId: 680636414
Change-Id: Ia7fd3fc738a2dd68b1063dba0dc847c5b0668401