Support %workspace% in execution log paths
Previously the execution logs were always written to relative paths.
This caused an issue if you put this flag in your `.bazelrc` and
expected the log to be produced in a relative directory. You can now use
`%workspace%` to handle writing to the root of your workspace.

Fixes https://github.com/bazelbuild/bazel/issues/22267