Simplify TargetRecorder API for disabling name conflict checking
Name conflict checking is disabled for package deserialization but enabled in all other contexts where we are building a `Package` object. Previously, this disabling was done by a stateful mutation on the `Package.Builder`, with associated precondition checks to confirm that this mutation did or did not occur before calling certain methods, and that this mutation was only attempted at most once.

This is a bit convoluted and is the result of my original attempt to understand `Package.Builder`'s sprawling API (now part of `TargetRecorder`). It's clear though that this can be better expressed as a constructor arg, removing the need for the extra consistency checking.

Also updated some javadoc to remove mention of fields being nullified after package construction -- that was removed in https://github.com/bazelbuild/bazel/commit/89778912d0eb038cbe0e7d9978103b9f7772b899. The field `macroNamespaceViolatingTargets` is also now never null -- even during package deserialization it still eventually gets a value due to the `putAll...()` method, so we may as well initialize it to a collection object. With these changes, the fields in question are now able to be marked `final`.

Drive-by cleanup during #19922.

PiperOrigin-RevId: 681065954
Change-Id: I060ae2fbaca9589ed95a45568127538a6aafe537