Add --run_in_client option to mobile-install
This option allows the mobile-install deployer command to be sent to the bazel client for execution. This is useful for configurations where the bazel client is on a different machine than the bazel server.

Takes the shortcut of having --run_in_client not respect
--mobile_install_run_deployer. The latter (undocumented) flag exists as an
earlier attempt at something like --run_in_client and AFAICT is only used for
that purpose, so taking advantage of that situation to avoid introducing extra
flags. We make these two flags get along after an internal compatibility window
has passed.

PiperOrigin-RevId: 681104164
Change-Id: Id462debabc54766b4b4f86024e4378f76798fd74