Introduce `FilesetOutputTree`.
This is a mechanical refactoring to replace `ImmutableList<FilesetOutputSymlink>` with an extra abstraction layer called `FilesetOutputTree`.

In this change, `FilesetOutputTree` only contains `ImmutableList<FilesetOutputSymlink>`. In a subsequent change, we will also store a `boolean` indicating whether the fileset has any relative symlinks. If the fileset does not have any relative symlinks (expected to be a common case), we can make some significant performance optimizations.

PiperOrigin-RevId: 681116734
Change-Id: If49f8acf15122f6287cd426cb36a081952424b84