Java part of python nostamp mitigation support.
The purpose of this change is to modify py_binary building so that when binaries are built with the --nostamp flag, most stamp info is still suppressed, but the build _target_ is still stamped.

This change will bring py_binary targets into line with current behaviour in C++, Java, Go and Rust.

PiperOrigin-RevId: 682299016
Change-Id: If33314ca3daf6c9b68daca3e539f7493119ddebf