[7.4.0] Support %workspace% in execution log paths (#23877)
Previously the execution logs were always written to relative paths.
This caused an issue if you put this flag in your `.bazelrc` and
expected the log to be produced in a relative directory. You can now use
`%workspace%` to handle writing to the root of your workspace.

Fixes https://github.com/bazelbuild/bazel/issues/22267

Closes #23813.

PiperOrigin-RevId: 682196916
Change-Id: I5f6fe3c72cfc3fe9a891e0d44440cb49bbc787d8

Commit
https://github.com/bazelbuild/bazel/commit/9aaeb9fe2c1edaca38f91c53a5620a2df2220302

Co-authored-by: Keith Smiley <keithbsmiley@gmail.com>