Propagate all experimental options to the exec configuration
Experimental flags that don't propagate to the exec configuration can create very surprising failure cases for both Bazel users and devs. Unless there is a valid reason to not propagate them, they should be.

This is now enforced by a test and existing flags have been cleaned up where possible. The test strategy is the same as for `incompatible_` flags.

Closes #23326.

PiperOrigin-RevId: 682447129
Change-Id: I18d6f70301388afbab34884a813de8ac86dae3b7