Remove rules_proto dependency from java_tools
This is a blocker for removal of proto rules. Because of this dependency, not everything is loaded from rules repositories and java_tools's cc_proto_library in particular falls back to native rule. See: https://buildkite.com/bazel/google-bazel-presubmit/builds/84956

After the change we need java_tools and rules_java release.

PiperOrigin-RevId: 683066764
Change-Id: Ic978dbb3058d60510f04021a278b6896f26bb754