Fix NPE in execlog when runfiles middleman are not top-level
The cherry-pick of the compact execlog runfiles changes assumed that runfiles only appear in the top-level tool nested set, but this isn't true for genrules. Instead, as in Bazel 8, it can be assumed that all middleman are also contained in inputs, which makes it possible to remove the extra middleman parameter and pass all runfiles tree information down to transitive log calls.