Use a visitor pattern to consume fileset symlinks.
Instead of constructing a `FilesetManifest`, callers use a `Visitor` that is invoked for every symlink in the fileset. This allows for an optimization: when the fileset has no relative symlinks (or we are ignoring relative symlinks), we can do a single-pass visitation without building up maps.

Additionally, callers are responsible for concatenating the fileset-relative and execroot-relative paths to their roots. This reduces garbage from concatenating paths, since some callers don't need them.

PiperOrigin-RevId: 683222046
Change-Id: I80f04306e7a617d3c3e630c56d96b2c968eb1d42