Automated rollback of commit 8b19650b5817a80fce115b3ede20e92a409eabd6.
*** Reason for rollback ***

After the root cause for b/356561371 has been identified, and the fix unknown commit is in, this blunt mitigation is no longer needed.

Independent of this, investigations and considerations on how to prevent this in the future are still ongoing.

*** Original change description ***

Blunt mitigation for blaze crashing with includes_scanning error: Adding retries for the 2 error modes.

The as yet not understood root cause of this crash must be an unhandled race condition where a *.blaze-grepped_includes_CPP file is read before it is present and fully available.
The two fixes are:
1. Waiting for the file to exist if it doesn't. This wait is very cheap, and it covers the majority of error cases.
2. Retry reading the file should an IOError occur while reading the file. This is...

***

PiperOrigin-RevId: 683997262
Change-Id: I6d1f612b1e7ffa146afec8d517e5f28145f12241