Deserializer for FileInvalidationData.
The core of FileDependencyDeserializer is an asynchronous state machine
implemented using Futures.transformAsync to recursively add more stages
as needed. Asynchrony is a consequence of the fact that the resolution
mechanism is an asynchronous remote cache lookup for FileInvalidationData
and deserialization typically requires multiple, sequential resolution
steps.

It uses the usual concurrent map to deduplicate keyed computations. In
this case, the concurrent map is a weak-valued Caffeine instance. The
rationale is that the map values (which are effectively sets of files)
are referenced by the SkyValues that would be invalidated by changes
to those files. So as the SkyValues are invalidated and GCd, the cache
is automatically cleaned up. This fits well with the transitive structure
of the FileDependencies class.

## Servicification Considerations

The cache keys are versioned so it is theoretically possible to share
a single cache across multiple Bazel instances.

At the moment, the BUILD dependencies are somewhat broad due to
including Skyframe serialization, but only the FingerprintValueService
is needed. Otherwise coupling to Bazel code is deliberately minimized.

PiperOrigin-RevId: 684437123
Change-Id: Iec445c222d6f072bd10044102dbcb01c62ab739d